Task #3208: Final merge from the trunk before this branch can be reintegrated.

+++ branches/CEP-Pipeline-Task3208/MAC/APL/RTCCommon/include/APL/RTCCommon/NsTimestamp.h
@@ -91,7 +91,7 @@ public:
 	// Return parts.
 	long sec()  const;
 	long nsec() const;
-	static NsTimestamp maxTime() { return (NsTimestamp(LONG_MAX, 999999999)); };
+	static NsTimestamp maxTime() { return (NsTimestamp(INT_MAX, 999999999)); };
 	/*@}*/
 
 	/*@{*/
